sereal: Move to own pkg
